requirements.txt
asgiref==3.5.2
Django==3.2
django-autoslug==1.9.8
django-ckeditor==6.3.0
django-js-asset==2.0.0
Pillow==8.3.0
pytz==2022.1
sqlparse==0.4.2
